Elevplaner/requirements.txt
lxml==4.6.5
python-docx==0.8.11
